convert some file to packages for commonly using
