+++ contracts/cw-bitcoin/src/lib.rs
@@ -2,12 +2,10 @@ mod app;
 mod checkpoint;
 mod constants;
 pub mod entrypoints;
-pub mod error;
 mod header;
 mod interface;
 pub mod msg;
 
-mod adapter;
 pub mod contract;
 mod fee;
 pub mod helper;
