+++ contracts/cw-bitcoin/src/signatory.rs
@@ -2,13 +2,11 @@ use std::cmp::Ordering;
 
 use crate::app::ConsensusKey;
 use crate::constants::MAX_SIGNATORIES;
-use crate::interface::Xpub;
 use crate::state::get_validators;
 use crate::state::SIG_KEYS;
 use crate::state::XPUBS;
+use common_bitcoin::xpub::Xpub;
 
-use super::error::ContractError;
-use super::error::ContractResult;
 use super::threshold_sig::Pubkey;
 use bitcoin::blockdata::opcodes::all::{
     OP_ADD, OP_CHECKSIG, OP_DROP, OP_ELSE, OP_ENDIF, OP_GREATERTHAN, OP_IF, OP_SWAP,
@@ -17,6 +15,8 @@ use bitcoin::blockdata::opcodes::{self, OP_FALSE};
 use bitcoin::blockdata::script::{read_scriptint, Instruction};
 use bitcoin::Script;
 use bitcoin_script::bitcoin_script as script;
+use common_bitcoin::error::ContractError;
+use common_bitcoin::error::ContractResult;
 use cosmwasm_schema::schemars::JsonSchema;
 use cosmwasm_schema::serde::{Deserialize, Serialize};
 use cosmwasm_std::Order;
