+++ packages/common-bitcoin/Cargo.toml
@@ -1,10 +1,12 @@
 [package]
 authors = {workspace = true}
+description = "Common bitcoin bridge types and forward types"
 documentation = {workspace = true}
 edition = {workspace = true}
 homepage = {workspace = true}
 license = {workspace = true}
-name = "cw-bitcoin"
+name = "common-bitcoin"
+readme = "README.md"
 repository = {workspace = true}
 version = {workspace = true}
 
@@ -14,39 +16,24 @@ crate-type = ["cdylib", "rlib"]
 
 [features]
 # use library feature to disable all instantiate/execute/query exports
-default = ['mainnet']
-library = []
-mainnet = []
-test-tube = []
+default = []
 
 [dependencies]
 arrayref = "0.3"
 bitcoin = {workspace = true, features = ["serde"]}
 bitcoin-script = {workspace = true}
-common-bitcoin = {workspace = true}
 cosmwasm-schema = {workspace = true}
 cosmwasm-std = {workspace = true}
 cw-storage-plus = {workspace = true}
-cw2 = {workspace = true}
-cw20 = {workspace = true}
-cw20-base = {workspace = true}
 derive_more = {workspace = true}
 libsecp256k1-core = "0.3.0"
-oraiswap = {workspace = true}
-prost = {workspace = true, features = ["prost-derive"]}
-sha2 = {workspace = true}
-skip = {workspace = true}
 thiserror = {workspace = true}
-token-bindings = {workspace = true}
-tokenfactory = {workspace = true, features = ["library"]}
 
 [dev-dependencies]
 bitcoincore-rpc-async = {workspace = true}
 bitcoind = {workspace = true}
-cosmwasm-testing-util = {workspace = true}
 serde = {workspace = true}
 serde_json = {workspace = true}
-tokio = {workspace = true}
 
 chrono = "0.4.19"
 serial_test = "3.1.1"
