+++ packages/common-bitcoin/src/adapter.rs
