+++ packages/common-bitcoin/src/error.rs
