+++ packages/common-bitcoin/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod adapter;
+pub mod error;
+pub mod xpub;
